samples: net: echo-server: Refactor tunnel support
Move the common tunnel setup code to samples/net/common directory
so that other network samples can utilize that too.
