+++ samples/net/common/Kconfig
@@ -38,3 +38,15 @@ config NET_SAMPLE_COMMON_VLAN_SETUP_2
 	  The VLAN tag can have values between 0 and 4094.
 	  There can be multiple IP addresses specified, and the
 	  code tries to apply them all to this interface.
+
+config NET_SAMPLE_COMMON_TUNNEL_PEER_ADDR
+	string "Remote IP address of the tunnel interface"
+	depends on NET_L2_IPIP
+	help
+	  Use overlay-tunnel.conf to setup the tunnel support.
+
+config NET_SAMPLE_COMMON_TUNNEL_MY_ADDR
+	string "My address for tunnel interface"
+	depends on NET_L2_IPIP
+	help
+	  The value depends on your network setup.
