+++ samples/net/common/common.cmake
@@ -7,3 +7,6 @@ target_sources(app PRIVATE $ENV{ZEPHYR_BASE}/samples/net/common/net_sample_commo
 
 target_sources_ifdef(CONFIG_NET_VLAN app PRIVATE
   $ENV{ZEPHYR_BASE}/samples/net/common/vlan.c)
+
+target_sources_ifdef(CONFIG_NET_L2_IPIP app PRIVATE
+  $ENV{ZEPHYR_BASE}/samples/net/common/tunnel.c)
