+++ samples/net/common/net_sample_common.h
@@ -4,6 +4,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+#include <zephyr/net/net_if.h>
+
 #if defined(CONFIG_NET_CONNECTION_MANAGER)
 void wait_for_network(void);
 #else
@@ -18,3 +20,19 @@ static inline int init_vlan(void)
 	return 0;
 }
 #endif /* CONFIG_NET_VLAN */
+
+#if defined(CONFIG_NET_L2_IPIP)
+int init_tunnel(void);
+bool is_tunnel(struct net_if *iface);
+#else
+static inline int init_tunnel(void)
+{
+	return 0;
+}
+
+static inline bool is_tunnel(struct net_if *iface)
+{
+	ARG_UNUSED(iface);
+	return false;
+}
+#endif /* CONFIG_NET_L2_IPIP */
