+++ samples/net/sockets/echo_server/CMakeLists.txt
@@ -21,7 +21,6 @@ endif()
 target_sources(                     app PRIVATE src/echo-server.c)
 target_sources_ifdef(CONFIG_NET_UDP app PRIVATE src/udp.c)
 target_sources_ifdef(CONFIG_NET_TCP app PRIVATE src/tcp.c)
-target_sources_ifdef(CONFIG_NET_L2_IPIP app PRIVATE src/tunnel.c)
 
 if (CONFIG_USB_DEVICE_STACK)
   target_sources(app PRIVATE src/usb.c)
