+++ samples/net/sockets/echo_server/Kconfig
@@ -27,18 +27,6 @@ config NET_SAMPLE_IFACE2_MY_IPV4_NETMASK
 	help
 	  The value depends on your network setup.
 
-config NET_SAMPLE_TUNNEL_PEER_ADDR
-	string "Remote IP address of the tunnel interface"
-	depends on NET_L2_IPIP
-	help
-	  Use overlay-tunnel.conf to setup the tunnel support.
-
-config NET_SAMPLE_TUNNEL_MY_ADDR
-	string "My address for tunnel interface"
-	depends on NET_L2_IPIP
-	help
-	  The value depends on your network setup.
-
 config NET_SAMPLE_PSK_HEADER_FILE
 	string "Header file containing PSK"
 	default "dummy_psk.h"
