+++ samples/net/sockets/echo_server/overlay-tunnel.conf
@@ -6,11 +6,11 @@ CONFIG_NET_IF_MAX_IPV6_COUNT=4
 CONFIG_NET_IF_MAX_IPV4_COUNT=4
 
 # First tunneling interface will have these settings
-CONFIG_NET_SAMPLE_TUNNEL_PEER_ADDR="192.0.2.2"
-#CONFIG_NET_SAMPLE_TUNNEL_PEER_ADDR="2001:db8::2"
+CONFIG_NET_SAMPLE_COMMON_TUNNEL_PEER_ADDR="192.0.2.2"
+#CONFIG_NET_SAMPLE_COMMON_TUNNEL_PEER_ADDR="2001:db8::2"
 
 # Use this for IPv6 over IPv4
-CONFIG_NET_SAMPLE_TUNNEL_MY_ADDR="2001:db8:200::1"
+CONFIG_NET_SAMPLE_COMMON_TUNNEL_MY_ADDR="2001:db8:200::1"
 
 # Use this for IPv4 over IPv4
-#CONFIG_NET_SAMPLE_TUNNEL_MY_ADDR="198.51.100.1"
+#CONFIG_NET_SAMPLE_COMMON_TUNNEL_MY_ADDR="198.51.100.1"
